+++ R/rc_sim.R
@@ -5,7 +5,6 @@
 #' @param soil_properties (data.table) Data table with soil properties: A_C_OF, soil organic carbon content (g/kg) or B_C_ST03, soil organic carbon stock (Mg C/ha), preferably for soil depth 0.3 m; A_CLAY_MI, clay content (\%); A_DENSITY_SA, dry soil bulk density (g/cm3)
 #' @param A_DEPTH (numeric) Depth for which soil sample is taken (m). Default set to 0.3.
 #' @param B_DEPTH (numeric) Depth of the cultivated soil layer (m), simulation depth. Default set to 0.3.
-#' @param M_TILLAGE_SYSTEM (character) gives the tillage system applied. Options include NT (no-till), ST (shallow-till), CT (conventional-till) and DT (deep-till).
 #' @param rothc_rotation (data.table) Table with crop rotation details and crop management actions that have been taken. Includes also crop inputs for carbon. See details for desired format.
 #' @param rothc_amendment (data.table) A table with the following column names: P_DATE_FERTILIZATION, P_HC, and B_C_OF_INPUT and/or P_DOSE and P_C_OF. See details for desired format.
 #' @param rothc_parms (list) A list with simulation parameters controlling the dynamics of RothC Model. For more information, see details.
@@ -43,12 +42,12 @@
 #'
 #' rothc_parms: parameters to adapt calculations (optional)
 #' May include the following columns:
-#' * initialize (boolean) scenario to initialize the carbon pools. Options TRUE or FALSE, default is TRUE
+#' * initialization_method (character) scenario to initialize the carbon pools. options 'spinup_analytical_bodemcoolstof','spinup_analytical_heuvelink', 'spinup_simulation', 'none', default is 'spinup_analytical_bodemcoolstof'
 #' * c_fractions (list) Distribution over the different C pools. If not supplied nor calculated via model initialization, default RothC distribution is used
 #' * dec_rates (list) list of decomposition rates of the different pools. If not supplied, default RothC values are used
 #' * unit (character) Unit in which the output should be given. Options: 'A_SOM_LOI' (\% organic matter),'psoc' (g C/kg), 'psomperfraction' (\% organic matter of each fraction), 'cstock' (kg C/ha of each fraction)
 #' * method (character) method to solve ordinary differential equations, see \link[deSolve]{ode} for options. default is adams.
-#' * poutput (character) Resolution of data ouptut. Options: 'year', 'month'
+#' * poutput (character) Resolution of data output. Options: 'year', 'month'
 #' * start_date (character, formatted "YYYY-MM-DD") Start date of simulation period. If not provided, first date of crop rotation or amendment application is taken.
 #' * end_date (character, formatted "YYYY-MM-DD") End date of simulation period. If not provided, last date of crop rotation or amendment application is taken.
 #' 
@@ -75,7 +74,6 @@
 rc_sim <- function(soil_properties,
                    A_DEPTH = 0.3,
                    B_DEPTH = 0.3,
-                   M_TILLAGE_SYSTEM = 'CT',
                    rothc_rotation = NULL,
                    rothc_amendment = NULL,
                    rothc_parms = NULL,
@@ -123,7 +121,7 @@ rc_sim <- function(soil_properties,
   poutput <- rothc_parms$poutput
   
   # Define initialize
-  initialize <- rothc_parms$initialize
+  initialization_method <- rothc_parms$initialization_method
   
   # Define dates of complete simulation period
   dt.time <- rc_time_period(start_date = start_date, end_date = end_date)
@@ -152,7 +150,6 @@ rc_sim <- function(soil_properties,
     dt.org <- NULL
   }
 
-
   # make rate modifying factors input database
   if(!is.null(dt.crop)){
   dt.rmf <- rc_input_rmf(dt = dt.crop,
@@ -170,6 +167,7 @@ rc_sim <- function(soil_properties,
                            dt.irrigation = irrigation)
     
   }
+ 
   # combine RothC input parameters
   rothc.parms <- list(k1 = k1,k2 = k2, k3=k3, k4=k4, R1 = dt.rmf$R1, abc = dt.rmf$abc, time = dt.rmf$time)
   
@@ -182,77 +180,46 @@ rc_sim <- function(soil_properties,
   # prepare EVENT database with all C inputs over time 
   rothc.event <- rc_input_events(crops = event.crop,amendment = event.man)
   
-  # initialize the RothC pools (kg C / ha)
-  
   # make internal data.table 
   dt.soc <- data.table(A_C_OF = soil_properties$A_C_OF,B_C_ST03 = soil_properties$B_C_ST03, A_CLAY_MI = soil_properties$A_CLAY_MI,a_depth = A_DEPTH,b_depth = B_DEPTH, A_DENSITY_SA = soil_properties$A_DENSITY_SA)
   
   # Correct A_C_OF for sampling depth 
   dt.soc[a_depth < 0.3 & A_CLAY_MI <= 10, A_C_OF := A_C_OF * (1 - 0.19 * ((0.20 - (pmax(0.10, a_depth) - 0.10))/ 0.20))]
   dt.soc[a_depth < 0.3 & A_CLAY_MI > 10, A_C_OF := A_C_OF * (1 - 0.33 * ((0.20 - (pmax(0.10, a_depth) - 0.10))/ 0.20))]
-  
+ 
   # calculate total organic carbon (kg C / ha)
   if(length(dt.soc$B_C_ST03) != 0) {
     dt.soc[,toc := B_C_ST03 * 1000]
   }else{
     dt.soc[,toc := A_C_OF / 1000 * A_DENSITY_SA * 1000 * B_DEPTH * 100 * 100]
   }
- 
-  # set the default initialisation to the one used in BodemCoolstof
-  if(initialize == TRUE){
+
+  # initialize the RothC pools if requested 
+  if(initialization_method != 'none'){ 
     
-    # set TOC to ton C / ha
-    dt.soc[, toc := toc * 0.001]
+    # derive the initial distribution of C pools (original data.tables are used as input)
+    c_fractions = as.list(rc_initialise(crops = rothc_rotation, 
+                                        amendment = rothc_amendment,
+                                        dt.time = dt.time,
+                                        dt.soc = dt.soc,
+                                        soil_properties = soil_properties,
+                                        rothc.event = rothc.event,
+                                        start_date = start_date,
+                                        dt.weather = dt.weather,
+                                        rothc.parms = rothc.parms,
+                                        initialization_method = initialization_method))
     
-    # time correction (is 12 in documentation Chantals' study, not clear why, probably due to fixed time step calculation)
-    timecor = 1
-    
-    # set rate modifying parameters
-    abc <- rothc.parms$abc
-    
-    # CDPM pool (ton C / ha)
-    cdpm.ini <- rothc.event[var == 'CDPM',list(time,value)]
-    cdpm.ini[,cf_abc := abc(time)]
-    cdpm.ini <- cdpm.ini[,((sum(value) * 0.001 / max(time)) / (mean(cf_abc)/timecor))/k1]
-    dt.soc[, cdpm.ini := mean(cdpm.ini)]
-    
-    # CRPM pool (ton C / ha)
-    crpm.ini = rothc.event[var == 'CRPM',list(time,value)]
-    crpm.ini[,cf_abc := abc(time)]
-    crpm.ini <- crpm.ini[,((sum(value) * 0.001 / max(time)) / (mean(cf_abc)/timecor))/k2]
-    dt.soc[, crpm.ini := mean(crpm.ini)]
-    
-    # CIOM pool (ton C / ha)
-    dt.soc[, ciom.ini := 0.049 * toc^1.139]
-    
-    # CBIOHUM pool (ton C /ha)
-    dt.soc[,biohum.ini := toc - ciom.ini - crpm.ini - cdpm.ini]
-    
-    # set to defaults when RPM and DPM inputs exceeds 70% / 50% of total C to avoid negative values for initial C pools
-    dt.soc[biohum.ini <0, cdpm.ini := 0.015 * (toc-ciom.ini)]
-    dt.soc[biohum.ini <0, crpm.ini := 0.125 * (toc-ciom.ini)]
-    dt.soc[, biohum.ini := toc-ciom.ini - crpm.ini - cdpm.ini]
-    
-    # CBIO and CHUM pool
-    dt.soc[,cbio.ini := biohum.ini / (1 + k3 / k4)]
-    dt.soc[,chum.ini := biohum.ini / (1 + k4 / k3)]
-    
-    # Set the intial C pools (kg C / ha)
-    dt.soc[,CIOM0 := ciom.ini * 1000]
-    dt.soc[,CDPM0 := cdpm.ini * 1000]
-    dt.soc[,CRPM0 := crpm.ini * 1000]
-    dt.soc[,CBIO0 := cbio.ini * 1000]
-    dt.soc[,CHUM0 := chum.ini * 1000]
-  } else {
-    # Calculate carbon pools based on provided or default distribution (kg C / ha)
-    dt.soc[,CIOM0 := c_fractions$fr_IOM * ((toc*0.001)^1.139) * 1000]
-    dt.soc[,CDPM0 := c_fractions$fr_DPM * (toc-CIOM0)]
-    dt.soc[,CRPM0 := c_fractions$fr_RPM * (toc-CIOM0)]
-    dt.soc[,CBIO0 := c_fractions$fr_BIO * (toc-CIOM0)]
-    dt.soc[,CHUM0 := toc-CIOM0-CDPM0-CRPM0-CBIO0]
-    
-  }
   
+  } 
+
+# calculate initial pool sizes based on initialized or supplied fraction distribution (kg C / ha)
+dt.soc[,CIOM0 := c_fractions$fr_IOM * toc]
+dt.soc[,CDPM0 := c_fractions$fr_DPM * toc]
+dt.soc[,CRPM0 := c_fractions$fr_RPM * toc]
+dt.soc[,CBIO0 := c_fractions$fr_BIO * toc]
+dt.soc[,CHUM0 := pmax(0, toc-CIOM0-CDPM0-CRPM0-CBIO0)]
+ 
+
   # extract relevant columns
   rothc.ini <- dt.soc[,list(CIOM0,CDPM0,CRPM0,CBIO0,CHUM0)]
   
@@ -290,8 +257,7 @@ rc_sim <- function(soil_properties,
   # estimate total SOC (kg C/ha)
   out[,soc := round(CDPM + CRPM + CBIO + CHUM + dt.soc$CIOM0)]
   
- 
- 
+
   # select type output
   if(unit=='A_SOM_LOI') {
     # Output in organic matter content [\%]